requirements.txt
pandas>=2.0.0
scikit-learn>=1.0.0
plotly>=5.0.0
matplotlib>=3.0.0
seaborn>=0.11.0
pytest>=7.0.0
pytest-cov>=4.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
scipy>=1.0.0